[base] Pull request #1485: Bugfix/LINUXDAR-3799 only let Updatescheduler remove marker file when suldownloader has finished and not when it is restarting due to update
Merge in LINUXEP/everest-base from bugfix/LINUXDAR-3799-health-can-handle-upgrades-and-downgrades to develop

Squashed commit of the following:

commit 25975f127409e84a69396a5b7b23ac2b79cbff1d
Author: Menaga Karunanandan <Menaga.Karunanandan@sophos.com>
Date:   Fri Dec 10 11:19:48 2021 +0000

    [LINUXDAR-3799] fix typos

commit 03b122db6e22cf3ab647d5510c28c94d9ac454be
Author: Menaga Karunanandan <Menaga.Karunanandan@sophos.com>
Date:   Fri Dec 10 11:12:56 2021 +0000

    [LINUXDAR-3799] only let Updatescheduler remove marker file when suldownloader has finished and not when it is restarting due to update

+++ base/modules/SulDownloader/SULDownloader.cpp
@@ -426,7 +426,22 @@ namespace SulDownloader
         timeTracker.setFinishedTime(TimeUtils::getCurrTime());
         if (productChanging)
         {
-            fs->removeFile(Common::ApplicationConfiguration::applicationPathManager().getUpdateMarkerFile());
+            try
+            {
+                std::string path = Common::ApplicationConfiguration::applicationPathManager().getUpdateMarkerFile();
+                if (fs->isFile(path))
+                {
+                    fs->removeFile(Common::ApplicationConfiguration::applicationPathManager().getUpdateMarkerFile());
+                }
+                else
+                {
+                    LOGWARN("Marker file removed before update finished");
+                }
+            }
+            catch (Common::FileSystem::IFileSystemException& ex)
+            {
+                LOGWARN("Failed to remove marker file with error: " << ex.what());
+            }
         }
 
         // if any error happened during installation, it reports correctly.
